ettaselnasel/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
alibabacloud-dysmsapi20170525==2.0.24
alibabacloud-tea==0.3.9
alibabacloud_credentials==0.3.5
alibabacloud_endpoint_util==0.0.3
alibabacloud_gateway_spi==0.0.2
alibabacloud_openapi_util==0.2.2
alibabacloud_tea_openapi==0.3.11
alibabacloud_tea_util==0.3.13
alibabacloud_tea_xml==0.0.2
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
backports.zoneinfo==0.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
black==24.4.2
bleach==6.1.0
blinker==1.8.2
blis==0.7.11
catalogue==1.0.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==3.4.8
cymem==2.0.8
decorator==5.1.1
distlib==0.3.8
Django==4.2.14
dnspython==2.6.1
elastic-transport==8.15.0
elasticsearch==8.15.0
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz#sha256=06c80936324012d1223291d2af41a5229e746dc2dee8fe31a532666ee3d18aaa
filelock==3.15.4
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-CKEditor==1.0.0
Flask-Dance==7.1.0
Flask-Elasticsearch==0.2.5
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-OAuthlib==0.9.4
Flask-SMS==0.0.24
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.1
future==1.0.0
geocoder==1.38.1
greenlet==3.0.3
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
mysqlclient==2.2.4
npm==0.1.1
numpy==1.24.4
oauthlib==3.2.2
optional-django==0.1.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
pipenv==2024.0.1
plac==1.1.3
platformdirs==4.2.2
preshed==3.0.9
pycparser==2.22
ratelim==0.1.6
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
soupsieve==2.6
spacy==2.3.7
SQLAlchemy==2.0.31
sqlparse==0.5.1
srsly==1.0.7
thinc==7.4.6
threadpoolctl==3.5.0
tomli==2.0.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
URLObject==2.4.3
virtualenv==20.26.3
wasabi==0.10.1
webencodings==0.5.1
Werkzeug==3.0.3
Whoosh==2.7.4
WTForms==3.1.2
WTForms-SQLAlchemy==0.4.1
yarl==1.9.11
zipp==3.19.2
